move expand and load yaml to eth2.0-utils
