+++ packages/eth2.0-spec-test-util/package.json
@@ -40,6 +40,7 @@
     "blockchain"
   ],
   "dependencies": {
+    "@chainsafe/eth2.0-utils": "^0.0.1",
     "@chainsafe/ssz": "^0.5.1",
     "bn.js": "^5.0.0",
     "camelcase": "^5.3.1",
