+++ packages/eth2.0-spec-test-util/test/e2e/multi/index.test.ts
@@ -1,13 +1,13 @@
 import {IBaseCase, describeMultiSpec} from "../../../src";
 import path from "path";
 
-interface BulkTestCase extends IBaseCase {
+interface IBulkTestCase extends IBaseCase {
   input: string;
   output: string;
 }
 
-describeMultiSpec<BulkTestCase, string>(
-  path.join(__dirname, '../_test_files/multi/bulk.yml'),
+describeMultiSpec<IBulkTestCase, string>(
+  path.join(__dirname, "../_test_files/multi/bulk.yml"),
   (input) => input,
   (testCase => [testCase.input]),
   (testCase => testCase.output),
