+++ packages/eth2.0-spec-test-util/test/e2e/single/index.test.ts
@@ -2,10 +2,8 @@ import {describeDirectorySpecTest} from "../../../src/single";
 import {join} from "path";
 import {AnyContainerType, AnySSZType, serialize} from "@chainsafe/ssz";
 import {bool, number64} from "@chainsafe/eth2.0-ssz-types/lib/generators/primitive";
-import {readFileSync, writeFileSync, unlinkSync} from "fs";
-import {objectToCamelCase} from "../../../src/util";
-import {load} from "js-yaml";
-import {schema} from "../../../src/yaml/schema";
+import {unlinkSync, writeFileSync} from "fs";
+import {loadYamlFile} from "@chainsafe/eth2.0-utils";
 
 export interface ISimpleStruct {
   test: boolean;
@@ -66,17 +64,7 @@ describeDirectorySpecTest<ISimpleCase, number>(
   }
 );
 
-function yamlToSSZ(file: string, sszSchema: AnySSZType) {
-  const input = objectToCamelCase(
-    load(
-      readFileSync(
-        file,
-        'utf8'
-      ),
-      {
-        schema,
-      }
-    )
-  );
-  writeFileSync(file.replace('.yaml', '.ssz'), serialize(input, sszSchema));
+function yamlToSSZ(file: string, sszSchema: AnySSZType): void {
+  const input = loadYamlFile(file);
+  writeFileSync(file.replace(".yaml", ".ssz"), serialize(input, sszSchema));
 }
