+++ packages/eth2.0-utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./crypto";
 export * from "./merkleTree";
 export * from "./math";
+export * from "./yaml";
